ARM64: preserve insect capture actor addresses

+++ src/actor/ac_ins_hitodama.c
@@ -296,7 +296,7 @@ static void aIHD_setupAction(aINS_INSECT_ACTOR* insect, int action, GAME* game) 
 
 static void aIHD_actor_move(ACTOR* actorx, GAME* game) {
     aINS_INSECT_ACTOR* insect = (aINS_INSECT_ACTOR*)actorx;
-    u32 label;
+    uintptr_t label;
 
     aIHD_light_proc(actorx, game);
     if (mPlib_get_player_actor_main_index(game) != mPlayer_INDEX_PUTAWAY_NET) {
@@ -304,7 +304,7 @@ static void aIHD_actor_move(ACTOR* actorx, GAME* game) {
     }
 
     label = mPlib_Get_item_net_catch_label();
-    if (label == (u32)actorx) {
+    if (label == (uintptr_t)actorx) {
         aIHD_setupAction(insect, aIHD_ACTION_LET_ESCAPE, game);
     } else {
         if (mEv_check_status(mEv_EVENT_GHOST, mEv_STATUS_RUN) == FALSE) {
